requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.1.1
decorator==5.2.1
Django==5.2.4
Events==0.5
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
fsspec==2025.7.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.47.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
open3d==0.19.0
opencv-python==4.12.0.88
opensearch-py==3.0.0
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.1
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.17
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0